+++ host_win32/mini/mjsrt.cpp
@@ -86,7 +86,9 @@ static MObject *CopyObjectFromJsValue(JsValueRef value)
             }
             else
             {
-                return MIntCreate((int)raw);
+                //the double value may exceed the range of int.
+                //convert to int64_t first, so that the low bits can be got.
+                return MIntCreate((int)(int64_t)raw);
             }
         }
         case JsString:
